Performance work

+++ vsta/src/include/sys/hat.h
@@ -57,6 +57,6 @@ extern void hat_fork(struct vas *, struct vas *);
  * hat_getbits()
  *	Get bits from mapping, clear bits from HAT layer simultaneously
  */
-extern int hat_getbits(struct pview *, void *);
+extern uchar hat_getbits(struct pview *, void *);
 
 #endif /* _HAT_H */
